[ComicList] should not call service.removeComicData() when no comic select

+++ src/app/comic-downloader/comic-list/comic-list.component.ts
@@ -26,7 +26,8 @@ export class ComicListComponent implements OnInit {
   }
 
   removeComicData() {
-    this.service.removeComicData({name: '', url: this.selectedComic});
+    if (this.selectedComic) {
+      this.service.removeComicData({ name: '', url: this.selectedComic });
+    }
   }
-
 }
